requirements.txt
# api-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
psycopg2-binary==2.9.5
redis==4.5.1
pydantic==1.10.7
kafka-python==2.0.2
pyspark==3.4.0